Kernel parameters handling: add python interface and refactor - move general utlities to "parameters_utils.h" - add python interface to avoid duplication of hashing code - add comments, docstrings - rename python script - remove obsolete options and functions - fix includes, add compilation rules
